% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4NC.R
\name{BS4NC}
\alias{BS4NC}
\title{BS4NC}
\usage{
BS4NC(dat, threshold, param, distr = "distr")
}
\description{
Brier Score.
This is implemented according to Equation 1 of the "Cross validation study and verification document"
}

