% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_NC.R
\name{create_empty_gofnc}
\alias{create_empty_gofnc}
\title{create_empty_gofnc}
\usage{
create_empty_gofnc(flood_database.nc_path = "output/flood_database.nc")
}
\description{
Function to create an empty gof_nc database
}

