% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_DATABASE.R
\name{create_empty_nc}
\alias{create_empty_nc}
\title{create_empty_nc}
\usage{
create_empty_nc(dat = postproc_dat, meta_dat = flood_metadata)
}
\arguments{
\item{dat}{Flood data. The default is the data inculded in this package.}
}
\description{
Function to create the structure of the NetCDF file.
}
\examples{
summary(flood_data)
# update_nc_structure()
}

