% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_DATABASE.R
\name{fillup_nc_parallel}
\alias{fillup_nc_parallel}
\title{fillup_nc_parallel}
\usage{
fillup_nc_parallel(dat = flood_data, meta_dat = flood_metadata,
  nc_path = "output/flood_database.nc")
}
\description{
Function to fill upthe NetCDF database
}

