% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_ALL.R
\name{fit_distr}
\alias{fit_distr}
\title{fit_distr}
\usage{
fit_distr(dat, distr = "distr", method = "method")
}
\arguments{
\item{method}{}
}
\description{
The function that calls the right fitting function for each distribution and fitting method
}

