% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4NC.R
\name{gof_ks}
\alias{gof_ks}
\title{gof_ks}
\usage{
gof_ks(dat, param, distr = "distr", test.stat = TRUE, p.value = FALSE)
}
\arguments{
\item{p.value}{}
}
\description{
Kolmogorov Smirnov. Gives results with warnings but this is because some flood data are repeated (for some strange reason)
}

