% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAD.R
\name{load_1sample}
\alias{load_1sample}
\title{load_1sample}
\usage{
load_1sample(dat.fulll, k, method = "method")
}
\arguments{
\item{method}{}
}
\description{
Function that returns only 1 sample with k omited values, according to a specific method (random, sorted min and max)
}

