% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_4_REPORT.R
\name{norway_map}
\alias{norway_map}
\title{norway_map}
\usage{
norway_map(dat, station.nb.vect)
}
\arguments{
\item{station.nb.vect}{}
}
\description{
Figure: MAP OF NORWAY WITH ALL THE STATIONS. WORKS
}

