% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_4_REPORT.R
\name{norway_map2}
\alias{norway_map2}
\title{norway_map2}
\usage{
norway_map2(dat, station.nb.vect)
}
\arguments{
\item{station.nb.vect}{}
}
\description{
MAP OF NORWA WITH COLOR CHANGING ACCORDING TO THE NUMBER OF DATA.
}

