% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot_all}
\alias{plot_all}
\title{plot_all}
\usage{
plot_all(dat, GOF.list, param, distr = "distr", method = "method")
}
\arguments{
\item{method}{}
}
\description{
This function creates a 2 by 2 plot by calling all common plotting functions
It plots for any distribution and method,
as well as the distribution parameters and Goodness Of Fit values
Returns nothing, saves nothing
}

