% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(dat, GOF.list, param, distr = "distr")
}
\arguments{
\item{distr}{}
}
\description{
plot fitted probability density function to estimated empirical pdf
Returns nothing, saves nothing
}

