% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{plot_ecdf}
\usage{
plot_ecdf(dat, param, distr = "distr")
}
\arguments{
\item{distr}{}
}
\value{
Returns nothing, saves nothing
}
\description{
Plot estimated and empirical cumulative distribution function
}

