% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot_qq}
\alias{plot_qq}
\title{plot_qq}
\usage{
plot_qq(dat, param, distr = "distr")
}
\arguments{
\item{distr}{}
}
\value{
Returns nothing, saves nothing
}
\description{
QQ plot of empiricial against modelled
}

