% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS.R
\name{plot_rlevel}
\alias{plot_rlevel}
\title{plot_rlevel}
\usage{
plot_rlevel(dat, param, distr = "distr")
}
\arguments{
\item{distr}{}
}
\value{
Returns nothing, saves nothing
}
\description{
Plot return levels
}

