% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4NC.R
\name{QS4NC}
\alias{QS4NC}
\title{QS4NC}
\usage{
QS4NC(dat, r.levels, r.periods)
}
\description{
Quantile Score. Only uses dat.sample so the calling script should be sending DAT.SAMPLE!
This is implemented according to Equation 2 of the "Cross validation study and verification document"
}

