% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4NC.R
\name{run_all_ff}
\alias{run_all_ff}
\title{run_all_ff}
\usage{
run_all_ff(dat, station.nb.vect, distr = "distr", method = "method")
}
\arguments{
\item{method}{}
}
\description{
FF function copied from run_all.r
Test of run_all for ff only.
}

