% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_ALL.R
\name{run_all_ff2}
\alias{run_all_ff2}
\title{run_all_ff2}
\usage{
run_all_ff2(dat, station.nb.vect, distr = "distr", method = "method")
}
\description{
Test of run_all for ff only.
TO COMPARE WITH run_all_ff in F4NCE.R!!!
}

