% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_ALL.R
\name{run_all_norm}
\alias{run_all_norm}
\title{run_all_norm}
\usage{
run_all_norm(dat, station.nb.vect, distr = "distr", method = "method")
}
\arguments{
\item{method}{}
}
\description{
The function to run all distributions and all methods and output csv files of the GOF
}

