% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAD.R
\name{save_coordinates}
\alias{save_coordinates}
\title{save_coordinates}
\usage{
save_coordinates(dat, station.nb.vect)
}
\arguments{
\item{station.nb.vect}{}
}
\description{
This function saves a new csv with the utm coordinates of each station in station.nb.vect
}

